lxc: adjusted libvirt-lxc process to add restore mode to it.
When doing lxc migration or simply restoring the container from a
saved state, we need restore the container from CRIU img files that
we have stored in disk. In this patch, we should extend
lxcContainerStart into a more generic one, that either starts a container
from scratch or restores it from a snapshot.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>